Updated Utils and moved some
